Socket: fix unix socket reuse problem
Issue solution: https://github.com/nginx/unit/issues/669.

Earlier, @VBart suggested to implement a check in case of ECONNREFUSED.

The problem happened a little earlier for bind() and threw an EADDRINUSE error.

It was also suggested by @mar0x to check if the file type matches the socket before deleting it.

Discussion:
https://github.com/nginx/unit/pull/649